Move delayheap to lib package
